# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가 시스템 공통기능 API
  version: 1.0.0
  description: 사용자·조직·역할·권한·메뉴·공통코드의 1차 범위 HTTP 계약.
servers:
- url: /api
  description: 브라우저 상대경로 API
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: 사용자 목록 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-001
      - REQ-002
      - REQ-028
      - REQ-032
      - REQ-045
      - REQ-050
      x-roles: &id001
      - R09
      parameters:
      - name: personnelNo
        in: query
        schema:
          type: string
        required: false
      - name: name
        in: query
        schema:
          type: string
        required: false
      - name: organization
        in: query
        schema:
          type: string
        required: false
      - name: position
        in: query
        schema:
          type: string
        required: false
      - name: employmentStatus
        in: query
        schema:
          type: string
        required: false
      - name: roleCode
        in: query
        schema:
          type: string
        required: false
      - name: useYn
        in: query
        schema:
          type: string
        required: false
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
        required: false
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
        required: false
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{userId}/settings:
    patch:
      operationId: updateUserSettings
      summary: 사용자 사용여부·역할 변경
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-003
      - REQ-028
      - REQ-029
      - REQ-030
      - REQ-041
      - REQ-033
      x-roles: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsRequest'
      x-business-rules:
      - KORUS 원천 인사 정보는 변경하지 않는다
      - use_yn과 업무 역할만 내부 관리 대상으로 변경한다
      x-side-effects:
      - user_account.use_yn과 user_role 매핑을 저장한다
      - change_history에 변경 전후 값·처리자·처리일시·사유를 기록한다
      x-required-tests:
      - 'happy: R09가 useYn=Y와 roleCodes를 저장 -> user_account와 user_role을 재조회하면 변경값이 보인다'
      - 'auth: R09가 아닌 세션 PATCH -> 403 ApiError를 반환한다'
      - 'validation: useYn에 Y/N 이외 값을 전송 -> useYn 필드 오류 ApiError를 반환한다'
      - 'business: KORUS 원천 personnel_no 변경 요청 -> korus_personnel_snapshot 값이 변하지 않고 요청이 차단된다'
      - 'side-effect: 저장 성공 -> change_history에 user_account 전후 값과 처리자가 존재한다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /users/{userId}/roles:
    get:
      operationId: listUserRoles
      summary: 사용자 역할 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-010
      x-roles: *id001
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
        required: false
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
        required: false
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveUserRole
      summary: 사용자 역할 부여
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-011
      - REQ-012
      - REQ-029
      - REQ-030
      - REQ-041
      - REQ-033
      x-roles: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleRequest'
      x-business-rules:
      - 사용자에게 하나 이상의 역할을 부여·변경·회수할 수 있다
      - 승인자와 유효기간을 기록한다
      x-side-effects:
      - user_role을 저장 또는 상태 변경한다
      - change_history에 사용자 역할 변경을 기록한다
      x-required-tests:
      - 'happy: R09가 R01 역할과 approvalUserId를 부여 -> user_role 재조회에 유효기간과 승인자가 보인다'
      - 'auth: R09가 아닌 세션 POST -> 403 ApiError를 반환한다'
      - 'validation: approvalUserId가 없는 요청 -> approvalUserId 필드 오류 ApiError를 반환한다'
      - 'business: 정의되지 않은 roleCode 부여 요청 -> user_role 행을 만들지 않고 오류를 반환한다'
      - 'side-effect: 부여 성공 -> change_history에 user_role 변경 이력이 존재한다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /code-groups:
    get:
      operationId: listCodeGroups
      summary: 코드그룹 목록 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-022
      - REQ-024
      x-roles: *id001
      parameters:
      - name: groupId
        in: query
        schema:
          type: string
        required: false
      - name: useYn
        in: query
        schema:
          type: string
        required: false
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
        required: false
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
        required: false
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveCodeGroup
      summary: 코드그룹 저장
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-023
      - REQ-029
      - REQ-030
      - REQ-041
      - REQ-042
      - REQ-033
      x-roles: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGroupRequest'
      x-business-rules:
      - 그룹ID·명칭·설명·관리부서를 관리한다
      x-side-effects:
      - code_group을 저장한다
      - change_history에 코드그룹 변경을 기록한다
      x-required-tests:
      - 'happy: R09가 groupId와 groupName을 저장 -> code_group 재조회에 해당 그룹이 보인다'
      - 'auth: R09가 아닌 세션 POST -> 403 ApiError를 반환한다'
      - 'validation: groupId가 없는 요청 -> groupId 필드 오류 ApiError를 반환한다'
      - 'business: 기존 groupId 중복 등록 -> 새 code_group 행을 만들지 않고 오류를 반환한다'
      - 'side-effect: 저장 성공 -> change_history에 code_group 변경 이력이 존재한다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /code-groups/{groupId}/detail-codes:
    get:
      operationId: listDetailCodes
      summary: 상세코드 목록 조회
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-025
      - REQ-027
      x-roles: *id001
      parameters:
      - name: useYn
        in: query
        schema:
          type: string
        required: false
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
        required: false
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
        required: false
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveDetailCode
      summary: 상세코드 저장
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-026
      - REQ-027
      - REQ-029
      - REQ-030
      - REQ-031
      - REQ-041
      - REQ-042
      - REQ-033
      x-roles: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetailCodeRequest'
      x-business-rules:
      - 코드값·코드명·상위코드·정렬순서·추가속성을 관리한다
      - 사용 중 상세코드는 물리삭제보다 use_yn 비활성화를 우선한다
      x-side-effects:
      - detail_code를 저장하고 상위코드 연결을 유지한다
      - change_history에 상세코드 변경을 기록한다
      x-required-tests:
      - 'happy: R09가 codeValue·parentDetailCodeId·additionalAttributes를 저장 -> detail_code 재조회에 계층과 속성이 보인다'
      - 'auth: R09가 아닌 세션 POST -> 403 ApiError를 반환한다'
      - 'validation: codeValue가 없는 요청 -> codeValue 필드 오류 ApiError를 반환한다'
      - 'business: useYn=N 상세코드 삭제 요청 -> detail_code 행은 남고 use_yn=N으로 재조회된다'
      - 'side-effect: 저장 성공 -> change_history에 detail_code 변경 이력이 존재한다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 입력 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 미인증
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: HttpOnly SameSite=Lax 세션 쿠키
  schemas:
    UseYn:
      type: string
      enum:
      - Y
      - N
    RoleCode:
      type: string
      enum:
      - R01
      - R02
      - R03
      - R04
      - R05
      - R06
      - R07
      - R08
      - R09
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
        meta:
          type: object
          additionalProperties: true
...
    UserSettingsRequest:
      type: object
      required:
      - useYn
      - roleCodes
      properties:
        useYn:
          $ref: '#/components/schemas/UseYn'
        roleCodes:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
        reason:
          type: string
...
    UserRoleRequest:
      type: object
      required:
      - roleCode
      - approvalUserId
      - effectiveStartDate
      properties:
        roleCode:
          $ref: '#/components/schemas/RoleCode'
        approvalUserId:
          type: string
        effectiveStartDate:
          type: string
          format: date
        effectiveEndDate:
          type: string
          format: date
        reason:
          type: string
...
    CodeGroupRequest:
      type: object
      required:
      - groupId
      - groupName
      properties:
        groupId:
          type: string
        groupName:
          type: string
        description:
          type: string
        managementDepartment:
          type: string
        useYn:
          $ref: '#/components/schemas/UseYn'
        reason:
          type: string
    DetailCodeRequest:
      type: object
      required:
      - codeValue
      - codeName
      - displayOrder
      properties:
        codeValue:
          type: string
        codeName:
          type: string
        parentDetailCodeId:
          type: string
          nullable: true
        displayOrder:
          type: integer
        additionalAttributes:
          type: object
          additionalProperties: true
        useYn:
          $ref: '#/components/schemas/UseYn'
        reason:
          type: string
...